TFP-4440 opplysninger om innleggelse i VKY-oppgave
